LCD menu optimalization - conversion functions - source separation
